# api/docs/swagger.yaml
definitions:
...
  models.CreateStaff:
    properties:
      balance:
        type: number
      branch_id:
        type: string
      name:
        type: string
      password:
        type: string
      tarif_id:
        type: string
      type:
        type: integer
      username:
        type: string
    type: object
...
  models.CreateStaffTransaction:
    properties:
      amount:
        type: number
      saleId:
        type: string
      sourceType:
        description: sales, bonus
        type: string
      staffId:
        type: string
      text:
        type: string
      type:
        description: withdraw, topup
        type: integer
    type: object
...
  models.GetAllStaffTransaction:
    properties:
      count:
        type: integer
      staffTransactions:
        items:
          $ref: '#/definitions/models.StaffTransaction'
        type: array
    type: object
...
  models.StaffTransaction:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      id:
        type: string
      saleId:
        type: string
      sourceType:
        description: sales, bonus
        type: string
      staffId:
        type: string
      text:
        type: string
      type:
        description: withdraw, topup
        type: integer
    type: object
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /staffTransaction:
    get:
      consumes:
      - application/json
      description: get staffTransaction
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        description: page of req
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllStaffTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: List staffTransactions
      tags:
      - staffTransactions
    post:
      consumes:
      - application/json
      description: api for create staffTransaction
      parameters:
      - description: date of staffTransaction
        in: body
        name: staffTransaction
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: create staffTransaction
      tags:
      - staffTransactions
  /staffTransaction/{id}:
    delete:
      consumes:
      - application/json
      description: api fot delete staffTransaction
      parameters:
      - description: id of staffTransaction
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: strig
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: delete staffTransaction
      tags:
      - staffTransactions
    get:
      consumes:
      - application/json
      description: get staffTransaction
      parameters:
      - description: id of staffTransaction
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: get staffTransaction
      tags:
      - staffTransactions
    put:
      consumes:
      - application/json
      description: api fot update staffTransaction
      parameters:
      - description: id of staffTransaction
        in: path
        name: id
        required: true
        type: string
      - description: id of staffTransaction
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: strig
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: updateda staffTransaction
      tags:
      - staffTransactions
swagger: "2.0"